feat: update sidebar with new component names and folders
