+++ integration-test/destination-async.spec.ts
@@ -18,7 +18,7 @@ export function handleAsyncDestinationTest() {
 
   test.describe.serial("Async destination", () => {
     test("should warn wrong resource ID", async ({ page }) => {
-      await page.goto("/destinations/create", { waitUntil: "networkidle" });
+      await page.goto("/data/create", { waitUntil: "networkidle" });
 
       // Should input destination id
       const idField = page.locator("input#destination-id");
@@ -46,7 +46,7 @@ export function handleAsyncDestinationTest() {
     });
 
     test("should create async destination", async ({ page }) => {
-      await page.goto("/destinations/create", { waitUntil: "networkidle" });
+      await page.goto("/data/create", { waitUntil: "networkidle" });
 
       // Should input destination id
       const idField = page.locator("input#destination-id");
@@ -117,7 +117,7 @@ export function handleAsyncDestinationTest() {
       await setupButton.isEnabled();
 
       await Promise.all([
-        page.waitForURL(`${env("NEXT_PUBLIC_CONSOLE_BASE_URL")}/destinations`),
+        page.waitForURL(`${env("NEXT_PUBLIC_CONSOLE_BASE_URL")}/data`),
         setupButton.click(),
       ]);
     });
@@ -125,7 +125,7 @@ export function handleAsyncDestinationTest() {
     test("should have destination list and navigate to destination details page", async ({
       page,
     }) => {
-      await page.goto("/destinations", { waitUntil: "networkidle" });
+      await page.goto("/data", { waitUntil: "networkidle" });
 
       // Should have model item in list
       const destinationItemTitle = page.locator("h3", {
@@ -136,14 +136,14 @@ export function handleAsyncDestinationTest() {
       // Should navigate to destination details page
       await Promise.all([
         page.waitForURL(
-          `${env("NEXT_PUBLIC_CONSOLE_BASE_URL")}/destinations/${destinationId}`
+          `${env("NEXT_PUBLIC_CONSOLE_BASE_URL")}/data/${destinationId}`
         ),
         page.locator("h3", { hasText: destinationId }).click(),
       ]);
     });
 
     test("should have proper destination details page", async ({ page }) => {
-      await page.goto(`/destinations/${destinationId}`, {
+      await page.goto(`/data/${destinationId}`, {
         waitUntil: "networkidle",
       });
 
@@ -205,7 +205,7 @@ export function handleAsyncDestinationTest() {
       const newS3OutputFormat = "JSON Lines: Newline-delimited JSON";
       const newS3OutputCompression = "GZIP";
 
-      await page.goto(`/destinations/${destinationId}`, {
+      await page.goto(`/data/${destinationId}`, {
         waitUntil: "networkidle",
       });
 
@@ -291,7 +291,7 @@ export function handleAsyncDestinationTest() {
       ]);
 
       // Reload page
-      await page.goto(`/destinations/${destinationId}`, {
+      await page.goto(`/data/${destinationId}`, {
         waitUntil: "networkidle",
       });
