+++ src/pages/data/create.tsx
@@ -1,23 +1,22 @@
 import { FC, ReactElement } from "react";
+import { GetServerSideProps } from "next";
 import { useRouter } from "next/router";
 import { shallow } from "zustand/shallow";
-
 import {
-  useWarnUnsavedChanges,
-  CreateSourceForm,
-  useCreateResourceFormStore,
   env,
+  useWarnUnsavedChanges,
+  CreateDestinationForm,
+  useCreateResourceFormStore,
   type CreateResourceFormStore,
 } from "@instill-ai/toolkit";
 
 import { PageTitle, PageHead, Topbar, Sidebar, PageBase } from "@/components";
-import { GetServerSideProps } from "next";
 
 export const getServerSideProps: GetServerSideProps = async () => {
   if (env("NEXT_PUBLIC_DISABLE_CREATE_UPDATE_DELETE_RESOURCE")) {
     return {
       redirect: {
-        destination: "/sources",
+        destination: "/data",
         permanent: false,
       },
     };
@@ -38,7 +37,7 @@ const selector = (state: CreateResourceFormStore) => ({
   init: state.init,
 });
 
-const CreateSourcePage: FC & {
+const CreateDestinationPage: FC & {
   getLayout?: FC<GetLayOutProps>;
 } = () => {
   const router = useRouter();
@@ -66,21 +65,21 @@ const CreateSourcePage: FC & {
 
   return (
     <>
-      <PageHead title="Create source-connector" />
+      <PageHead title="Create destination-connector" />
       <div className="flex flex-col">
         <PageTitle
-          title="Set Up New Source"
-          breadcrumbs={["Source", "Source Settings"]}
+          title="Set Up New Destination"
+          breadcrumbs={["Destination", "Destination Settings"]}
           disabledButton={true}
           marginBottom="mb-10"
         />
-        <CreateSourceForm
-          onCreate={(initStore) => {
+        <CreateDestinationForm
+          title={null}
+          onCreate={(_, initStore) => {
             initStore();
-            router.push("/sources");
+            router.push("/data");
           }}
           accessToken={null}
-          width="w-full"
           enabledQuery={true}
         />
       </div>
@@ -88,7 +87,7 @@ const CreateSourcePage: FC & {
   );
 };
 
-CreateSourcePage.getLayout = (page) => {
+CreateDestinationPage.getLayout = (page) => {
   return (
     <PageBase>
       <Topbar />
@@ -100,4 +99,4 @@ CreateSourcePage.getLayout = (page) => {
   );
 };
 
-export default CreateSourcePage;
+export default CreateDestinationPage;
