+++ src/pages/pipelines/create.tsx
@@ -8,7 +8,7 @@ export const getServerSideProps: GetServerSideProps = async () => {
   if (env("NEXT_PUBLIC_DISABLE_CREATE_UPDATE_DELETE_RESOURCE")) {
     return {
       redirect: {
-        destination: "/sources",
+        destination: "/operators",
         permanent: false,
       },
     };
